requirements.txt
Flask
pymongo
requests
beautifulsoup4
python-dotenv
Flask-Mail
Pillow
imap-tools